refactor: rename backend client to artists backend client and update references
